refactored main.py, implemented App class and its derivations
